% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_answered_questions.R
\name{all_answered_questions}
\alias{all_answered_questions}
\title{all_answered_questions}
\usage{
all_answered_questions(mp_id = NULL)
}
\arguments{
\item{mp_id}{Accepts a member ID, and returns a data frame with all available questions asked by that member. If NULL, returns  a data frame with all available answered questions.}
}
\description{
Imports data on answered parliamentary questions
}
\examples{
\dontrun{

x <- all_answered_questions(172)

}
}
\keyword{Answered}
\keyword{Questions}
