% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_members.R
\name{all_members}
\alias{all_members}
\title{All Members of Parliament}
\usage{
all_members(all = TRUE)
}
\arguments{
\item{all}{Imports all available members Defaults to TRUE.}
}
\description{
This imports data on All Members of Parliament including the Lords and the Commons
}
\examples{
all_members()
}
\keyword{All}
\keyword{Members}
\keyword{Parliament}
\keyword{of}

