% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bill_publication_types}
\alias{bill_publication_types}
\title{A character vector of possible types of bill publications. For use with \code{\link{bill_publications}}}
\format{A character vector containing 38 elements}
\usage{
bill_publication_types
}
\description{
A character vector of possible types of bill publications. For use with \code{\link{bill_publications}}
}
\seealso{
\code{\link{bill_publications}}

\code{\link{bills}}

\code{\link{bill_stage_types}}
}
\keyword{datasets}
