% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill_stage_types.R
\name{bill_stage_types}
\alias{bill_stage_types}
\alias{hansard_bill_stage_types}
\title{Bill Stage Types}
\usage{
bill_stage_types(tidy = TRUE, tidy_style = "snake_case", verbose = FALSE)

hansard_bill_stage_types(tidy = TRUE, tidy_style = "snake_case")
}
\arguments{
\item{tidy}{Fix the variable names in the tibble to remove special characters and superfluous text, and converts the variable names to a consistent style. Defaults to \code{TRUE}.}

\item{tidy_style}{The style to convert variable names to, if tidy = TRUE. Accepts one of 'snake_case', 'camelCase' and 'period.case'. Defaults to 'snake_case'.}

\item{verbose}{If \code{TRUE}, returns data to console on the progress of the API request. Defaults to \code{FALSE}.}
}
\description{
Returns a tibble with all possible bill stage types.
}
\examples{
\dontrun{
x <- bill_stage_types()
}
}
\seealso{
\code{\link{bills}}
}
