% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bills}
\alias{bills}
\title{Bills}
\usage{
bills(type = c("all", "ammended", "sponsor", "stage", "publications",
  "stage types"))
}
\arguments{
\item{all}{Imports all available bills Defaults to TRUE.}
}
\description{
This imports data on House of Commons and House of Lords bills
}
\examples{
bills()
}
\keyword{bills}

