% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_answered_questions.R
\name{commons_answered_questions}
\alias{commons_answered_questions}
\title{commons_answered_questions}
\usage{
commons_answered_questions(date = NULL, department = NULL,
  answered_by = NULL)
}
\arguments{
\item{date}{Returns a data frame with all answered questions in the House of Commons on the given date, in 'yyyy-mm-dd' format. Defaults to NULL.}

\item{department}{Returns a data frame with all answered questions in the House of Commons from the given department. Defaults to NULL.}

\item{answered_by}{Returns a data frame with all answered questions in the House of Commons by the given MP. Defaults to NULL.}

\item{all}{Returns a data frame with all answered questions in the House of Commons. The default response, but is overriden by other parameters. Defaults to TRUE.}
}
\description{
Imports data on House of Commons answered questions
}
\examples{
\dontrun{
x <- commons_answered_questions('all')

x <- commons_answered_questions('date')

x <- commons_answered_questions('department')

x <- commons_answered_questions('answered_by')
}
}
\keyword{bills}
