% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_bills.R
\name{commons_bills}
\alias{commons_bills}
\title{House of Commons Bills}
\usage{
commons_bills(all = TRUE)
}
\arguments{
\item{all}{Imports all available bills Defaults to TRUE.}
}
\description{
This imports data on House of Commons bills
}
\examples{
commons_bills()
}
\keyword{bills}

