% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_committee_inquiries-DEV.R
\name{commons_committee_inquiries}
\alias{commons_committee_inquiries}
\title{commons_committee_inquiries}
\usage{
commons_committee_inquiries(id = NULL, recommendations = FALSE, extra_args)
}
\arguments{
\item{id}{Returns the committee report of a given ID. If no ID number is entered, returns all available committee reports.}

\item{recommendations}{If TRUE, returns all reccomendations for the given report ID. Defaults to FALSE.}

\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}
}
\description{
Returns commons committee inquiries
}
\examples{
\dontrun{
x <- commons_committee_inquiries(id = NULL, recommendations = FALSE)
}
}
\keyword{Committee}
\keyword{Reports}
