% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_divisions.R
\name{commons_division_date}
\alias{commons_division_date}
\title{commons_division_date}
\usage{
commons_division_date(date = NULL, extra_args = NULL, tidy = TRUE)
}
\arguments{
\item{date}{Returns all divisions on a given date. Defaults to NULL.}

\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}

\item{tidy}{Fix the variable names in the tibble to remove extra characters, superfluous text and convert variable names to snake_case. Defaults to TRUE.}
}
\description{
Returns a tibble with the dates of House of Commons divisions.
}
\examples{
\dontrun{
x <- commons_division_date('2016-10-12')
}

}
\keyword{divisions}
