% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_interest.R
\name{commons_interest}
\alias{commons_interest}
\title{House of Commons Registered Interests}
\usage{
commons_interest(all = TRUE)
}
\arguments{
\item{all}{Imports all available House of Commons Registered Interests Defaults to TRUE.}
}
\description{
This imports data on House of Commons Registered Interests
}
\examples{
commons_interest()
}
\keyword{Commons}
\keyword{House}
\keyword{Interests}
\keyword{Registered}
\keyword{of}

