% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_members.R
\name{commons_members}
\alias{commons_members}
\title{House of Commons Members}
\usage{
commons_members(all = TRUE)
}
\arguments{
\item{all}{Imports all available House of Commons members Defaults to TRUE.}
}
\description{
This imports data on House of Commons members
}
\examples{
commons_members()
}
\keyword{Commons}
\keyword{House}
\keyword{Members}
\keyword{of}

