% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_oral_question_times.R
\name{commons_oral_question_times}
\alias{commons_oral_question_times}
\title{House of Commons Oral Question Times}
\usage{
commons_oral_question_times(all = TRUE)
}
\arguments{
\item{all}{Returns a data frame with all of the oral question times. Defaults to TRUE.}
}
\description{
Imports data on House of Commons oral question times
}
\examples{
\dontrun{
x <- commons_oral_question_times(all = TRUE)
}
}
\keyword{Oral}
\keyword{Questions}

