% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_oral_questions.R
\name{commons_oral_questions}
\alias{commons_oral_questions}
\title{commons_oral_questions}
\usage{
commons_oral_questions(comsOralType = c("all", "times", "daysTabled",
  "daysAnswered", "askedBy", "session", "department"))
}
\arguments{
\item{comsOralType}{The type of data you want, allows the arguments 'all', 'times', 'daysTabled', 'daysAnswered', 'askedBy' and 'session'}

\item{all}{Imports all available oral questions}

\item{times}{Imports the times of all available oral questions}

\item{daysTabled}{Requests two dates, and returns a data frame with all oral questions asked between those two dates}

\item{daysAnswered}{Requests two dates, and returns a data frame with all oral questions answered between those two dates}

\item{askedBy}{Requests a member ID and returns a data frame with all oral questions asked by that member}

\item{session}{Requests a session ID and returns a data frame with all oral questions asked in that session}

\item{department}{Requests a Requests a department name, and returns all oral questions answered by that department. The query acts as a search, so entering <health> will return all questions answered by the Department of Health}
}
\description{
Imports data on House of Commons oral questions
}
\examples{
\dontrun{
x <- commons_oral_questions('all')

x <- commons_oral_questions('times')

x <- commons_oral_questions('daysTabled')

x <- commons_oral_questions('daysAnswered')

x <- commons_oral_questions('askedBy')

x <- commons_oral_questions('session')

x <- commons_oral_questions('department')

}
}
\keyword{bills}
