% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_questions.R
\name{commons_questions}
\alias{commons_questions}
\title{House of Commons Oral Questions}
\usage{
commons_questions(type = c("all", "written", "oral"))
}
\arguments{
\item{all}{Imports all available oral questions. Defaults to TRUE.}
}
\description{
This imports data on House of Commons oral questiosn
}
\examples{
commons_questions()
}
\keyword{bills}

