% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_terms.R
\name{commons_terms}
\alias{commons_terms}
\title{commons_terms}
\usage{
commons_terms(extra_args = NULL)
}
\arguments{
\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}
}
\description{
Imports the parliamentary thesaurus.
}
\examples{
\dontrun{

x <- commons_terms(all)

}
}
\keyword{parliamentary}
\keyword{thesaurus}
