% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_console.R
\name{commons_written_questions_console}
\alias{commons_written_questions_console}
\title{commons_written_questions_console}
\usage{
commons_written_questions_console(comsWritType = c("all", "department",
  "dates", "askedBy"))
}
\arguments{
\item{comsWritType}{The type of data you want, allows the arguments 'all', 'department', 'dates' and 'askedBy'}

\item{all}{Returns a data frame with all written questions}

\item{department}{Requests a department name, and returns all written questions by department. The query acts as a search, so entering <health> will return all questions answered by the Department of Health}

\item{dates}{Requests two dates and imports a data frame with all available written questions from between the two dates.}

\item{askedBy}{Requests a member ID and returns a data frame with all written questions asked by that member}
}
\description{
Requests data on House of Commons written questions
}
\examples{
\dontrun{
x <- commons_written_questions_console('all')

x <- commons_written_questions_console('department')

x <- commons_written_questions_console('dates')

x <- commons_written_questions_console('askedBy')
}
}
\keyword{#'}
\keyword{@rdname}
\keyword{Commons}
\keyword{House}
\keyword{Questions}
\keyword{Written}
\keyword{hansard_console}
\keyword{internal}
\keyword{of}
