% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_day_motions.R
\name{early_day_motions}
\alias{early_day_motions}
\title{early_day_motions}
\usage{
early_day_motions(mp_id = NULL, edm_id = NULL, sponsor = NULL,
  all_sponsors = NULL, signatures = FALSE)
}
\arguments{
\item{mp_id}{Accepts a given MP and returns a data frame with all early day motions signed by that MP. Defaults to NULL.}

\item{sponsor}{Returns a data frame with all EDMs sponsored by a given MP. Defaults to NULL.}

\item{signatures}{Returns a data frame of all early day motion signatures.}

\item{primary_sponsor}{Requests a member's ID, and returns a data frame of early day motions where the given member is the primary sponsor}
}
\description{
Imports data on early day motions
}
\examples{
\dontrun{
x <- early_day_motions('all')

x <- early_day_motions('allSponsors')

x <- early_day_motions('all')

x <- early_day_motions('primarySponsor')

x <- early_day_motions('signatures')

x <- early_day_motions('ID')
}
}
\keyword{EDM}
