% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_day_motions.R
\name{edm_text}
\alias{edm_text}
\title{Early Day Motion Text}
\usage{
edm_text(id, tidy = TRUE, tidy_style = "snake", verbose = TRUE)
}
\arguments{
\item{id}{The ID of an individual Early Day Motion, or a vector of IDs,
as found in the \code{about} column of returns from \code{\link[=early_day_motions]{early_day_motions()}}}

\item{tidy}{Logical parameter. If \code{TRUE}, fixes the variable names
in the tibble to remove special characters and superfluous text, and
converts the variable names to a consistent style. Defaults to \code{TRUE}.}

\item{tidy_style}{The style to convert variable names to, if
\code{tidy = TRUE}. Accepts any style accepted by \link[snakecase]{to_any_case}.
Defaults to \code{'snake'}.}

\item{verbose}{If \code{TRUE}, displayes messages on the console on the
progress of the API request. Defaults to \code{TRUE}.}
}
\value{
A tibble of containing the EDM text and its ID.
}
\description{
A quick and dirty function for a specific use case, use with caution.
}
\examples{
\dontrun{
y <- edm_text(c("811291", "811292", "811293"))
}
}
