% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election.R
\name{elections}
\alias{elections}
\title{Elections}
\usage{
elections(all = TRUE)
}
\arguments{
\item{all}{Returns a data frame with the date and type of all general and by-elections since 1945. Defaults to TRUE.}
}
\description{
Imports data on elections
}
\examples{
\dontrun{
x <- elections(all = TRUE)
}
}
\keyword{Elections}

