% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epetition.R
\name{epetition}
\alias{epetition}
\title{ePetitions}
\usage{
epetition(type = c("all", "date", "no", "aye"))
}
\arguments{
\item{type}{The type of data you want, allows the arguments "all", "department" and "dates"}

\item{all}{Imports all available papers laid}
}
\description{
This imports data on petitions
}
\examples{
x <- papers_laid("all")

}
\keyword{Laid}
\keyword{Papers}

