\name{hansard}
\alias{hansard}
\title{Hansard}
\usage{
all_answered_questions()


all_members()


commons_answered_questions()


commons_bills()


commons_divisions()


commons_interest()


commons_members()


commons_oral_question_times()


commons_oral_questions()


commons_terms()


commons_terms()


commons_written_questions()


constituencies()


early_day_motions()


election_results()


elections()


lords_ammendments()


lords_attendance()


lords_divisions()


lords_interest()

lords_members()


lords_written_questions()


papers_laid()


petitions()


publication_logs()


research_briefings()


sessions()


tv_programmes()


}
\description{
Retrieves data from the data.parliament.uk API. Despite the name, the API does not current give access to the Hansard.
}


\examples{
all_answered_questions()


all_members()


commons_answered_questions()


commons_bills()


commons_divisions()


commons_interest()


commons_members()


commons_oral_question_times()


commons_oral_questions()


commons_terms()


commons_terms()


commons_written_questions()


constituencies()


early_day_motions()


election_results()


elections()


lords_ammendments()


lords_attendance()


lords_divisions()


lords_interest()

lords_members()


lords_written_questions()


papers_laid()


petitions()


publication_logs()


research_briefings()


sessions()


tv_programmes()

}
