% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_basic.R
\name{hansard_basic}
\alias{hansard_basic}
\title{Hansard - Generic API Function}
\usage{
hansard_basic()
}
\arguments{
\item{call}{Calls the}
}
\description{
Walks the user through the various steps of calling data from the parliamentary API. STILL A WORK IN PROGRESS, DO NOT USE!!!!
}
\examples{
\dontrun{
x <- hansard_basic('elections.json')
}
}
\keyword{API}
\keyword{Hansard}

