% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_generic.R
\name{hansard_generic}
\alias{hansard_generic}
\title{Hansard - Generic API Function}
\usage{
hansard_generic(path)
}
\arguments{
\item{path}{The url path to the data you wish to retrieve.}
}
\description{
A semi-generic function for the parliamentary API. Provides greater
flexibility, including any newly released features or data not yet
included in the individual functions of the hansard package.
}
\details{
Users must specify '.json?' in their path. The function uses the default
of 10 items per page, to include more include \code{'&_pageSize=[number]'},
e.g. \code{'&_pageSize=500'} to specifiy the maximum amount of
500 items per page.

This function does not tidy any variable names.
}
\examples{
\donttest{
x <- hansard_generic('elections.json?')

y <- hansard_generic('elections.json?electionType=General Election')
}
}
