% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_tidy.R
\name{hansard_tidy}
\alias{hansard_tidy}
\title{hansard_tidy}
\usage{
hansard_tidy(df, tidy_style)
}
\arguments{
\item{df}{The tibble to tidy}

\item{tidy_style}{The style to tidy variable names to.}
}
\value{
A tidied tibble
}
\description{
hansard_tidy
}
