% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_amendments.R
\name{lords_amendments}
\alias{lords_amendments}
\title{lords_amendments}
\usage{
lords_amendments(decision = NULL, extra_args = NULL)
}
\arguments{
\item{decision}{The decision on the amendements. Accepts one of "Withdrawn", "Agreed", "Disagreed", "Pending", "NotMoved", "Disposed".}

\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}
}
\description{
Imports data on House of Lords Amendments. Returns a data frame with all available House of Lords Amendments. Defaults to TRUE.
}
\examples{
\dontrun{
x <- lords_amendments()
}
}
\keyword{Amendments}
\keyword{House}
\keyword{Lords}
\keyword{of}
