% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard-defunct.R
\name{lords_ammendments}
\alias{lords_ammendments}
\title{Defunct parameters for hansard package}
\usage{
lords_ammendments(all = TRUE)
}
\arguments{
\item{all}{}
}
\value{

}
\description{
Defunct parameters for hansard package
}
\examples{

}
