% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_attendance.R
\name{lords_attendance}
\alias{lords_attendance}
\title{lords_attendance}
\usage{
lords_attendance(session_id = NULL)
}
\arguments{
\item{session_id}{The ID of the House of Lords session. If NULL, returns a list of all sessions. Defaults to NULL.}
}
\description{
Imports data on House of Lords attendance. Please note that the attendance data frames are not as tidy as some of the others that are accessible through this API.
}
\examples{
\dontrun{

x <- lords_attendance(session_id = 706178)
}

}
\keyword{Attendance}
\keyword{House}
\keyword{Lords}
\keyword{of}
