% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_divisions.R
\name{lords_divisions}
\alias{lords_divisions}
\title{House of Lords Divisions}
\usage{
lords_divisions(all = TRUE)
}
\arguments{
\item{all}{Imports all available House of Lords divisions Defaults to TRUE.}
}
\description{
This imports data on House of Lords divisions
}
\examples{
x <- lords_divisions()
}
\keyword{Divisions}
\keyword{Lords}

