% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_interest.R
\name{lords_interest}
\alias{lords_interest}
\title{House of Lords Registered Interests}
\usage{
lords_interest(all = TRUE)
}
\arguments{
\item{all}{Imports all available House of Lords registered interests Defaults to TRUE.}
}
\description{
This imports data on House of Lords registered interests
}
\examples{
lords_interest()
}
\keyword{House}
\keyword{Interests}
\keyword{Lords}
\keyword{Registered}
\keyword{of}

