% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_members.R
\name{lords_members}
\alias{lords_members}
\title{House of Lords Members}
\usage{
lords_members(all = TRUE)
}
\arguments{
\item{all}{Imports all available House of Lords members Defaults to TRUE.}
}
\description{
This imports data on House of Lords members
}
\examples{
lords_members()
}
\keyword{House}
\keyword{Lords}
\keyword{Members}
\keyword{of}

