% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members}
\alias{members}
\title{All Members of Parliament}
\usage{
members(house = c("all", "commons", "commons interests", "lords",
  "lords interests"))
}
\arguments{
\item{all}{Imports all available members Defaults to TRUE.}
}
\description{
This imports data on All Members of Parliament including the Lords and the Commons
}
\examples{
x <- members("all")
# Returns a data frame with information on all members of Parliament, including both the House of Lords and the House of Commons. The data frame includes both current and previous members, and the API currently does not have information on when a member first sat in the house, or to distinguish current from former members.

y <- members("commons")
# Returns a data frame with information on all members of the House of Commons. The data frame includes both current and previous members of the House of Commons, and the API currently does not have information on when a member first sat in the house, or to distinguish current from former members.

z <- members("lords")
# Returns a data frame with information on all members of the House of Lords. The data frame includes both current and previous members of the House of Lords, and the API currently does not have information on when a member first sat in the house, or to distinguish current from former members.

a <- members("commons interests") #Not Working
# Returns a list of data frames, each details on the registered interests of a member of the House of Commons

b <- members("lords interests")
# Returns a dataframe of the registered interests a member of the House of Lords, selected by their ID.
}
\keyword{All}
\keyword{Members}
\keyword{Parliament}
\keyword{of}

