% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{MP.ID}
\alias{MP.ID}
\title{MP ID}
\usage{
MP.ID(mpidSearch = NULL)
}
\arguments{
\item{all}{Imports all available members Defaults to TRUE.}
}
\description{
This imports data on All Members of Parliament including the Lords and the Commons
}
\examples{
MP.ID("query")







}
\keyword{All}
\keyword{Members}
\keyword{Parliament}
\keyword{of}

