% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_questions.R
\name{mp_questions}
\alias{mp_questions}
\title{mp_questions}
\usage{
mp_questions(mp_id, question_type = "all", extra_args = NULL)
}
\arguments{
\item{mp_id}{The ID number of a member of the House of Commons.}

\item{question_type}{Accepts the arguments 'all', 'oral' and 'written'. Defaults to 'all'.}

\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}
}
\description{
Accepts an ID number for a member of the House of Commons, and returns a data frame of of all their oral and written questions.
}
\examples{
\dontrun{
x <- mp_questions(172, 'all')
}

}
\keyword{questions}
