% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_briefings.R
\name{research_briefings}
\alias{research_briefings}
\title{Parliamentary Research Briefings}
\usage{
research_briefings(all = TRUE)
}
\arguments{
\item{all}{Imports all available Parliamentary Research Briefings Defaults to TRUE.}
}
\description{
This imports data on  Parliamentary Research Briefings
}
\examples{
x <- research_briefings()
}
\keyword{Briefings}
\keyword{Parliamentary}
\keyword{Research}

