% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_briefings_lists.R
\name{research_topics_list}
\alias{research_topics_list}
\alias{hansard_research_topics_list}
\alias{research_subtopics_list}
\alias{hansard_research_subtopics_list}
\alias{research_types_list}
\alias{hansard_research_types_list}
\title{Lists of research briefing topics, subtopics and types.}
\usage{
research_topics_list()

hansard_research_topics_list()

research_subtopics_list()

hansard_research_subtopics_list()

research_types_list()

hansard_research_types_list()
}
\value{
A list with the different research topics/subtopics/types available.
}
\description{
Returns lists of research briefing topics, subtopics and types. These functions do not accept any arguments.
}
\section{Member details functions}{

\describe{
\item{research_topics_list}{A list with the different research topics available}
\item{research_subtopics_list}{A list of Parliamentary Research Briefings topics}
\item{research_types_list}{A list of types of Parliamentary Research Briefings}
}
}

\examples{
 \dontrun{
research_topics_list <- research_topics_list()

research_subtopics_list <- research_subtopics_list()

research_types_list <- research_types_list()
}

}
\seealso{
\code{\link{research_briefings}}
}
