% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{research_topics_list}
\alias{research_topics_list}
\title{A list of parliamentary research briefing topics. The list may be used to support calls to the \code{\link{research_briefings}} API.}
\format{A list with twenty levels, one for each possible briefing topic.}
\source{
\url{http://explore.data.parliament.uk/}
}
\usage{
research_topics_list
}
\description{
A list of parliamentary research briefing topics. The list may be used to support calls to the \code{\link{research_briefings}} API.
}
\keyword{datasets}
