% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{research_types_list}
\alias{research_types_list}
\title{A list of parliamentary research briefing types.  The list may be used to support calls to the \code{\link{research_briefings}} API.}
\format{A list with six levels, one for each possible briefing type}
\source{
\url{http://explore.data.parliament.uk/}
}
\usage{
research_types_list
}
\description{
A list of parliamentary research briefing types.  The list may be used to support calls to the \code{\link{research_briefings}} API.
}
\keyword{datasets}
