% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions_info.R
\name{sessions_info}
\alias{sessions_info}
\title{Parliamentary Sessions}
\usage{
sessions_info(all = TRUE)
}
\arguments{
\item{all}{Imports Parliamentary Sessions Defaults to TRUE.}
}
\description{
This imports data on Parliamentary Sessions
}
\examples{
x <- sessions_info()
}
\keyword{Parliamentary}
\keyword{Sessions}

