% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_programmes.R
\name{tv_programmes}
\alias{tv_programmes}
\title{TV Programmes}
\usage{
tv_programmes(type = c("TV", "clips"))
}
\arguments{
\item{type}{The type of data you want, allows the arguments "TV" and "clips"}

\item{TV}{Imports data on all TV broadcasts produced by the Westminster Parliament}

\item{clips}{Imports data on all video clips produced by the Westminster Parliament}
}
\description{
This imports data on TV broadcasts
}
\examples{
x <- tv_programmes("TV")
# Returns a data frame with details on all TV broadcasts produced by the
# Westminster Parliament

x <- tv_programmes("clips")
# Returns a data frame with details on all video clips produced by the
# Westminster Parliament
}
\keyword{TV}

