% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_answered_questions.R
\name{all_answered_questions}
\alias{all_answered_questions}
\title{Parliamentary Answered Questions}
\usage{
all_answered_questions(all = TRUE)
}
\arguments{
\item{all}{Imports data on all available answered questions Defaults to TRUE.}
}
\description{
This imports data on answered parliamentary questions
}
\examples{
all_answered_questions()
}
\keyword{Answered}
\keyword{Questions}

