% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_answered_questions.R
\name{all_answered_questions}
\alias{all_answered_questions}
\title{all_answered_questions}
\usage{
all_answered_questions(mp_id = NULL, start_date = "1900-01-01",
  end_date = Sys.Date(), extra_args = NULL)
}
\arguments{
\item{mp_id}{Accepts a member ID, and returns a data frame with all available questions asked by that member. If NULL, returns  a data frame with all available answered questions.}

\item{start_date}{The earliest date to include in the data frame. Defaults to '1900-01-01'.}

\item{end_date}{The latest date to include in the data frame. Defaults to current system date.}

\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}
}
\description{
Imports data on answered parliamentary questions
}
\examples{
\dontrun{

x <- all_answered_questions(4019)

}
}
\keyword{Answered}
\keyword{Questions}
