% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_answered_questions.R
\name{all_answered_questions}
\alias{all_answered_questions}
\title{Parliamentary Answered Questions}
\usage{
all_answered_questions(allAnsType = c("all", "askedBy"))
}
\arguments{
\item{allAnsType}{The type of questions you want, accepts the arguments 'all' and 'askedBy'}

\item{all}{Returns a data frame with all available answered questions.}

\item{askedBy}{Requests a member ID, and returns a data frame with all available questions asked by that member.}
}
\description{
Imports data on answered parliamentary questions
}
\examples{
\dontrun{

x <- all_answered_questions('all')

x <- all_answered_questions('askedBy')
}
}
\keyword{Answered}
\keyword{Questions}

