% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill_stage_types.R
\name{bill_stage_types}
\alias{bill_stage_types}
\alias{hansard_bill_stage_types}
\title{Bill Stage Types}
\usage{
bill_stage_types(tidy = TRUE, tidy_style = "snake", verbose = TRUE)

hansard_bill_stage_types(tidy = TRUE, tidy_style = "snake", verbose = TRUE)
}
\arguments{
\item{tidy}{Logical parameter. If \code{TRUE}, fixes the variable names
in the tibble to remove special characters and superfluous text, and
converts the variable names to a consistent style. Defaults to \code{TRUE}.}

\item{tidy_style}{The style to convert variable names to, if
\code{tidy = TRUE}. Accepts any style accepted by \link[snakecase]{to_any_case}.
Defaults to \code{'snake'}.}

\item{verbose}{If \code{TRUE}, displayes messages on the console on the
progress of the API request. Defaults to \code{TRUE}.}
}
\value{
A tibble of bill stage types.
}
\description{
Returns a tibble with all possible bill stage types.
}
\examples{
\dontrun{
x <- bill_stage_types()
}

}
\seealso{
\code{\link[=bills]{bills()}}

\code{\link[=bill_publications]{bill_publications()}}
}
