% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bills}
\alias{bills}
\title{Bills}
\usage{
bills(billType = c("ammended", "stageTypes", "publications"))
}
\arguments{
\item{billType}{The type of data you want, allows the arguments "ammended", "publications" and "stageTypes"}

\item{ammended}{Imports a data frame with all ammended bills}

\item{stageTypes}{Imports a data frame with all bill stage types}

\item{publications}{Imports a data frame with all bill publications}
}
\description{
This imports data on House of Commons and House of Lords bills
}
\examples{
\donttest{
x <- bills("all")

x <- bills("ammended")

x <- bills("stageTypes")
}
}
\keyword{bills}

