% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bills}
\alias{bills}
\alias{hansard_bills}
\title{Bill data}
\usage{
bills(ID = NULL, amendments = FALSE, start_date = "1900-01-01",
  end_date = Sys.Date(), extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)

hansard_bills(ID = NULL, amendments = FALSE, start_date = "1900-01-01",
  end_date = Sys.Date(), extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)
}
\arguments{
\item{ID}{The ID of a given bill to return data on. If \code{NULL}, returns all bills, subject to other parameters. Defaults to \code{NULL}.}

\item{amendments}{If \code{TRUE}, returns all bills with amendments. Defaults to \code{FALSE}.}

\item{start_date}{The earliest date to include in the tibble. Defaults to \code{'1900-01-01'}. Accepts character values in \code{'YYYY-MM-DD'} format, and objects of class \code{Date}, \code{POSIXt}, \code{POSIXct}, \code{POSIXlt} or anything else than can be coerced to a date with \code{as.Date()}.}

\item{end_date}{The latest date to include in the tibble. Defaults to current system date. Defaults to \code{'1900-01-01'}. Accepts character values in \code{'YYYY-MM-DD'} format, and objects of class \code{Date}, \code{POSIXt}, \code{POSIXct}, \code{POSIXlt} or anything else than can be coerced to a date with \code{as.Date()}.}

\item{extra_args}{Additional parameters to pass to API. Defaults to \code{NULL}.}

\item{tidy}{Fix the variable names in the tibble to remove special characters and superfluous text, and converts the variable names to a consistent style. Defaults to \code{TRUE}.}

\item{tidy_style}{The style to convert variable names to, if \code{tidy = TRUE}. Accepts one of \code{'snake_case'}, \code{'camelCase'} and \code{'period.case'}. Defaults to \code{'snake_case'}.}

\item{verbose}{If \code{TRUE}, returns data to console on the progress of the API request. Defaults to \code{FALSE}.}
}
\value{
A tibble with details on bills before the House of Lords and the House of Commons.
}
\description{
Imports data on House of Commons and House of Lords bills
}
\examples{
\dontrun{

x <- bills()

x <- bills(amendments=TRUE)

x <- bills(1719)

x <- bills(start_date ='2016-01-01')

}
}
\seealso{
\code{\link{bill_stage_types}}
}
