% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bills}
\alias{bills}
\title{Bills}
\usage{
bills(type = c("all", "ammended", "sponsor", "stage", "publications",
  "stageTypes"))
}
\arguments{
\item{type}{The type of data you want, allows the arguments "all", "ammended", "sponsor", "stage", "publications", "stageTypes"}

\item{all}{Imports all available bills}

\item{ammended}{Imports all ammended bills}

\item{sponsor}{Imports bills by sponsor}

\item{stage}{Imports bills by stage}

\item{publications}{Imports data on publications}

\item{stageTypes}{Imports all stage types}
}
\description{
This imports data on House of Commons and House of Lords bills
}
\examples{
bills()
}
\keyword{bills}

