% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bills}
\alias{bills}
\title{Bills}
\usage{
bills(billType = c("ammended", "stageTypes", "publications"))
}
\arguments{
\item{billType}{The type of data you want, allows the arguments 'ammended', 'publications' and 'stageTypes'}

\item{ammended}{Imports a data frame with all ammended bills}

\item{stageTypes}{Imports a data frame with all bill stage types}

\item{publications}{Imports a data frame with all bill publications}
}
\description{
Imports data on House of Commons and House of Lords bills
}
\note{
There are problems with the Bills API, as the JSON data available for some queries,
including the query to return all bills currently before the house, is inconsistently formatted
and cannot be parsed into a data frame.
}
\examples{
\donttest{
#### NOT RUN:
# x <- bills('all')

# x <- bills('ammended')

# x <- bills('stageTypes')
}
}
\keyword{bills}

