% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_answered_questions.R
\name{commons_answered_questions}
\alias{commons_answered_questions}
\title{House of Commons Answered Questions}
\usage{
commons_answered_questions(type = c("all", "date", "department",
  "answered by", "recent"))
}
\arguments{
\item{all}{Imports all available answered questions Defaults to TRUE.}
}
\description{
This imports data on House of Commons answered questions
}
\examples{
commons_answered_questions()
}
\keyword{bills}

