% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_answered_questions.R
\name{commons_answered_questions}
\alias{commons_answered_questions}
\title{House of Commons Answered Questions}
\usage{
commons_answered_questions(comsAnsType = c("all", "date", "department",
  "answeredBy"))
}
\arguments{
\item{comsAnsType}{The type of data you want, allows the arguments 'all', 'date', 'department', 'answeredBy', 'recent'}

\item{all}{Returns a data frame with all answered questions in the House of Commons}

\item{date}{Returns a data frame with all answered questions in the House of Commons on the given date}

\item{department}{Returns a data frame with all answered questions in the House of Commons from the given department}

\item{answeredBy}{Returns a data frame with all answered questions in the House of Commons by the given MP}
}
\description{
Imports data on House of Commons answered questions
}
\examples{
\dontrun{
x <- commons_answered_questions('all')

x <- commons_answered_questions('date')

x <- commons_answered_questions('department')

x <- commons_answered_questions('answeredBy')
}
}
\keyword{bills}

