% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_answered_questions.R
\name{commons_answered_questions}
\alias{commons_answered_questions}
\title{House of Commons Answered Questions}
\usage{
commons_answered_questions(comsAnsType = c("all", "date", "department",
  "answeredBy", "recent"))
}
\arguments{
\item{comsAnsType}{The type of data you want, allows the arguments 'all', 'date', 'department', 'answeredBy', 'recent'}

\item{all}{Imports all available answered questions}

\item{date}{Imports all available answered questions on a particular date}

\item{department}{Imports all available answered questions by answering department}

\item{answeredBy}{Imports all available answered questions by answering MP}

\item{recent}{Imports all available answered questions - NOT WORKING!}
}
\description{
This imports data on House of Commons answered questions
}
\examples{
\donttest{
#### NOT RUN:
# x <- commons_answered_questions('all')
# Returns a data frame with all answered questions in the House of Commons

# x <- commons_answered_questions('date')
# Returns a data frame with all answered questions in the House of Commons on the given date

# x <- commons_answered_questions('department')
# Returns a data frame with all answered questions in the House of Commons from the given department

# x <- commons_answered_questions('answeredBy')
# Returns a data frame with all answered questions in the House of Commons by the given MP

# x <- commons_answered_questions('recent')
# Returns a data frame with all answered questions from the House of Commons
}
}
\keyword{bills}

