% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_divisions.R
\name{commons_division_date}
\alias{commons_division_date}
\title{commons_division_date}
\usage{
commons_division_date(date)
}
\arguments{
\item{date}{The ID number of a member of the House of Commons.}
}
\description{
Accepts an ID number for a member of the house of commons, and returns a data frame of all divisions where they voted aye.
}
\examples{
\dontrun{
x <- commons_division_date(2016-10-10)
}

}
\keyword{divisions}

