% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_divisions.R
\name{commons_division_date}
\alias{commons_division_date}
\title{commons_division_date}
\usage{
commons_division_date(date = NULL)
}
\arguments{
\item{date}{Returns all divisions on a given date. Defaults to NULL.}
}
\description{
Returns a data frames with
}
\examples{
\dontrun{
x <- commons_division_date('2016-10-12')
}

}
\keyword{divisions}
