% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_divisions.R
\name{commons_divisions}
\alias{commons_divisions}
\title{House of Commons Divisions}
\usage{
commons_divisions(type = c("all", "date", "no", "aye"))
}
\arguments{
\item{type}{The type of data you want, allows the arguments "all", "date", "no", "aye"}

\item{all}{Imports all available divisions. Defaults to TRUE.}
}
\description{
This imports data on House of Commons divisions
}
\examples{
x <- commons_divisions("all")
# Returns all divisions

x <- commons_divisions("date")
# Returns all divisions on a given date

x <- commons_divisions("no")
# Returns all divisions in which a given MP voted no

x <- commons_divisions("aye")
# Returns all divisions in which a given MP voted aye


}
\keyword{divisions}

