% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_divisions.R
\name{commons_divisions}
\alias{commons_divisions}
\title{House of Commons Divisions}
\usage{
commons_divisions(all = TRUE)
}
\arguments{
\item{all}{Imports all available divisions. Defaults to TRUE.}
}
\description{
This imports data on House of Commons divisions
}
\examples{
commons_divisions()
}
\keyword{divisions}

