% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_divisions.R
\name{commons_divisions}
\alias{commons_divisions}
\title{House of Commons Divisions}
\usage{
commons_divisions(type = c("all", "date", "no", "aye"))
}
\arguments{
\item{type}{The type of data you want, allows the arguments "all", "date", "no", "aye"}

\item{all}{Imports all available divisions. Defaults to TRUE.}
}
\description{
This imports data on House of Commons divisions
}
\examples{
commons_divisions("all")

commons_divisions("date")

commons_divisions("no")

commons_divisions("aye")


}
\keyword{divisions}

