% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_divisions.R
\name{commons_divisions}
\alias{commons_divisions}
\title{commons_divisions}
\usage{
commons_divisions(division_id = NULL, summary = FALSE)
}
\arguments{
\item{division_id}{The id of a particular vote. If empty, returns a data frame with information on all commons divisions. Defaults to NULL.}

\item{summary}{If TRUE, returns a small data frame summarising a division outcome. Has no effect if `division_id` is empty. Defaults to FALSE.}
}
\description{
Imports data on House of Commons divisions.
}
\examples{
\dontrun{

x <- commons_divisions(division_id = 694163, summary = FALSE)

}
}
\keyword{divisions}
