% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_oral_question_times.R
\name{commons_oral_question_times}
\alias{commons_oral_question_times}
\title{commons_oral_question_times}
\usage{
commons_oral_question_times(session = NULL, question_id = NULL)
}
\arguments{
\item{session}{Accepts a session in format yyyy/yy (e.g. 2016/17) and returns a data frame of all oral question times from that session}

\item{question_id}{Accepts a question time ID, and returns a data frame of that question time.}
}
\description{
Imports data on House of Commons oral question times
}
\examples{
\dontrun{

commons_oral_question_times <- function(session='2016/17', question_id='685697')

}
}
\keyword{Oral}
\keyword{Questions}
\keyword{Time}
