% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_oral_questions.R
\name{commons_oral_questions}
\alias{commons_oral_questions}
\title{commons_oral_questions}
\usage{
commons_oral_questions(mp_id = NULL, answering_department = NULL,
  start_date = "1900-01-01", end_date = Sys.Date())
}
\arguments{
\item{mp_id}{The ID of a given MP. Defaults to NULL.}

\item{answering_department}{The department that answers a question}

\item{start_date}{The earliest date to include in the data frame. Defaults to '1900-01-01'.}

\item{end_date}{The latest date to include in the data frame. Defaults to current system date.}
}
\description{
Imports data on House of Commons oral questions
}
\examples{
\dontrun{

x <- commons_oral_questions(mp_id=4019, answering_department = 'education',
start_date='1900-01-01', end_date=Sys.Date())

}
}
\keyword{questions}
