% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_oral_questions.R
\name{commons_oral_questions}
\alias{commons_oral_questions}
\title{House of Commons Oral Questions}
\usage{
commons_oral_questions(all = TRUE)
}
\arguments{
\item{all}{Imports all available oral questions. Defaults to TRUE.}
}
\description{
This imports data on House of Commons oral questiosn
}
\examples{
commons_oral_questions()
}
\keyword{bills}

