% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_oral_questions.R
\name{commons_oral_questions}
\alias{commons_oral_questions}
\title{House of Commons Oral Questions}
\usage{
commons_oral_questions(type = c("all", "times"))
}
\arguments{
\item{type}{The type of data you want, allows the arguments "all" and "times"}

\item{all}{Imports all available oral questions}

\item{times}{Imports the times of all available oral questions}
}
\description{
This imports data on House of Commons oral questions
}
\examples{
x <- commons_oral_questions()
}
\keyword{bills}

