% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_terms.R
\name{commons_terms}
\alias{commons_terms}
\title{commons_terms}
\usage{
commons_terms(all = TRUE)
}
\arguments{
\item{all}{Returns a data frame with all definitions in the parliamentary thesaurus}
}
\description{
Imports the parliamentary thesaurus
}
\examples{
\dontrun{

x <- commons_terms(all)

}
}
\keyword{parliamentary}
\keyword{thesaurus}
