% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_terms.R
\name{commons_terms}
\alias{commons_terms}
\title{Parliamentary Thesaurus}
\usage{
commons_terms(all = TRUE)
}
\arguments{
\item{all}{Imports information from the parliamentary thesaurus Defaults to TRUE.}
}
\description{
This imports the parliamentary thesaurus
}
\examples{
x <- commons_terms()
}
\keyword{parliamentary}
\keyword{thesaurus}

