% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_terms.R
\name{commons_terms}
\alias{commons_terms}
\title{Parliamentary Thesaurus}
\usage{
commons_terms(termsType = c("all"))
}
\arguments{
\item{termsType}{The type of data you want, allows the argument "all"}

\item{all}{Imports all information in the parliamentary thesaurus}
}
\description{
This imports the parliamentary thesaurus
}
\examples{
\donttest{
x <- commons_terms("all")
#NOT RUN
# x <- commons_terms("all")
# head(x)
# Returns a data frame with all definitions in the parliamentary thesaurus
 }
}
\keyword{parliamentary}
\keyword{thesaurus}

