% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_terms.R
\name{commons_terms}
\alias{commons_terms}
\title{commons_terms}
\usage{
commons_terms(termsType = c("all"))
}
\arguments{
\item{termsType}{The type of data you want, allows the argument 'all'}

\item{all}{Returns a data frame with all definitions in the parliamentary thesaurus}
}
\description{
Imports the parliamentary thesaurus
}
\examples{
\dontrun{
x <- commons_terms('all')
head(x)
 }
}
\keyword{parliamentary}
\keyword{thesaurus}
