% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_written_questions.R
\name{commons_written_questions}
\alias{commons_written_questions}
\title{House of Commons Bills Written Questions}
\usage{
commons_written_questions(all = TRUE)
}
\arguments{
\item{all}{Imports all available written questions Defaults to TRUE.}
}
\description{
This imports data on House of Commons written questions
}
\examples{
commons_written_questions()
}
\keyword{Questions}
\keyword{Written}

