% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_written_questions.R
\name{commons_written_questions}
\alias{commons_written_questions}
\title{House of Commons Bills Written Questions}
\usage{
commons_written_questions(comsWritType = c("all", "department", "dates"))
}
\arguments{
\item{comsWritType}{The type of data you want, allows the arguments 'all', 'department' and 'dates'}

\item{all}{Imports all available written questions}

\item{department}{Imports all available written questions answered by a given department}

\item{dates}{Imports all available written questions from between two given dates}
}
\description{
Imports data on House of Commons written questions
}
\examples{
\donttest{
#### NOT RUN:
# x <- commons_written_questions('all')
#Returns all written questions

# x <- commons_written_questions('department')
#Returns all written questions by department. The query acts as a search, so entering <health>
# will return all questions answered by the Department of Health


# x <- commons_written_questions('dates')
#Returns all written questions tabled between two chosen dates
}
}
\keyword{Commons}
\keyword{House}
\keyword{Questions}
\keyword{Written}
\keyword{of}

