% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constituencies.R
\name{constituencies}
\alias{constituencies}
\title{constituencies}
\usage{
constituencies(all = TRUE)
}
\arguments{
\item{all}{Returns a data frame of all constituencies. Defaults to TRUE.}
}
\description{
Imports data on House of Commons constituencies
}
\examples{
\dontrun{
x <- constituencies()
}

}
\keyword{Constituencies}
