% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constituencies.R
\name{constituencies}
\alias{constituencies}
\alias{hansard_constituencies}
\title{House of Commons constituencies}
\usage{
constituencies(current = TRUE, extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)

hansard_constituencies(current = TRUE, extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)
}
\arguments{
\item{current}{If \code{TRUE}, returns only current constituencies. If FALSE, returns all current and former constituencies. Defaults to \code{TRUE}.}

\item{extra_args}{Additional parameters to pass to API. Defaults to \code{NULL}.}

\item{tidy}{Fix the variable names in the tibble to remove special characters and superfluous text, and converts the variable names to a consistent style. Defaults to \code{TRUE}.}

\item{tidy_style}{The style to convert variable names to, if \code{tidy = TRUE}. Accepts one of 'snake_case', 'camelCase' and 'period.case'. Defaults to 'snake_case'.}

\item{verbose}{If \code{TRUE}, returns data to console on the progress of the API request. Defaults to \code{FALSE}.}
}
\value{
A tibble with details of Westminster constituencies.
}
\description{
Imports data on House of Commons constituencies, returning a tibble of all current and former Westminster constituencies.
}
\examples{
\dontrun{

x <- constituencies()

x <- constituencies(current = FALSE)

}

}
