% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constituencies.R
\name{constituencies}
\alias{constituencies}
\title{constituencies}
\usage{
constituencies(extra_args = NULL)
}
\arguments{
\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}
}
\description{
Imports data on House of Commons constituencies
}
\examples{
\dontrun{
x <- constituencies()
}

}
\keyword{Constituencies}
