% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constituencies.R
\name{constituencies}
\alias{constituencies}
\title{House of Commons Constituencies}
\usage{
constituencies(contType = c("all"))
}
\arguments{
\item{contType}{The type of data you want, allows the arguments 'all'}

\item{all}{Imports all available constituencies information Defaults to TRUE.}
}
\description{
Imports data on House of Commons constituencies
}
\examples{
\donttest{
### NOT RUN:
#x <- constituencies('all')
# Returns a data frame of all constituencies
}

}
\keyword{Constituencies}

