% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constituencies.R
\name{constituencies}
\alias{constituencies}
\title{constituencies}
\usage{
constituencies(current = TRUE, extra_args = NULL)
}
\arguments{
\item{current}{If TRUE, returns only current constituencies. If FALSE, returns all current and former constituencies. Due to the structure of the API, this function has to request all available data, and then subset that data.}

\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}
}
\description{
Imports data on House of Commons constituencies, returning a data frame of all current and former Westminster constituencies.
}
\examples{
\dontrun{

x <- constituencies()

x <- constituencies(current = FALSE)

}

}
\keyword{Constituencies}
