% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constituencies.R
\name{constituencies}
\alias{constituencies}
\title{House of Commons Constituencies}
\usage{
constituencies(all = TRUE)
}
\arguments{
\item{all}{Imports all available constituencies information Defaults to TRUE.}
}
\description{
This imports data on House of Commons constituencies
}
\examples{
constituencies()
}
\keyword{Constituencies}

