% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_day_motions.R
\name{early_day_motions}
\alias{early_day_motions}
\title{early_day_motions}
\usage{
early_day_motions(edm_id = NULL, start_date = "1900-01-01",
  end_date = Sys.Date(), signatures = 1, extra_args = NULL)
}
\arguments{
\item{edm_id}{Accepts a given MP and returns a data frame with all early day motions signed by that MP. Defaults to NULL.}

\item{start_date}{The earliest date to include in the data frame, if calling all early day motions. Defaults to '1900-01-01'.}

\item{end_date}{The latest date to include in the data frame, if calling all early day motions. Defaults to current system date.}

\item{signatures}{The minimum number of signatures required for inclusion in the data frame. Defaults to 1.}

\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}
}
\description{
Imports data on early day motions
}
\examples{
\dontrun{
x <- early_day_motions('all')

x <- early_day_motions('allSponsors')

x <- early_day_motions('all')

x <- early_day_motions('primarySponsor')

x <- early_day_motions('signatures')

x <- early_day_motions('ID')
}
}
\keyword{EDM}
