% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_day_motions.R
\name{early_day_motions}
\alias{early_day_motions}
\title{Early Day Motions}
\usage{
early_day_motions(all = TRUE)
}
\arguments{
\item{all}{Imports all available EDMs. Defaults to TRUE.}
}
\description{
This imports data on early day motions
}
\examples{
early_day_motions()
}
\keyword{EDM}

