% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_day_motions.R
\name{early_day_motions}
\alias{early_day_motions}
\title{Early Day Motions}
\usage{
early_day_motions(type = c("all", "signatures"))
}
\arguments{
\item{type}{The type of data you want, allows the arguments "all", "signatures"}

\item{all}{Imports all available EDMs.}

\item{signatures}{Imports all available signatures.}
}
\description{
This imports data on early day motions
}
\examples{
if (requireNamespace("jsonlite", quietly = TRUE)) {
early_day_motions("all")
#Returns a data frame of all early day motions


early_day_motions("signatures")
#Returns a data frame of all early day motions signatures
}
}
\keyword{EDM}

