% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_day_motions.R
\name{early_day_motions}
\alias{early_day_motions}
\title{Early Day Motions}
\usage{
early_day_motions(edmType = c("all"))
}
\arguments{
\item{edmType}{The type of data you want, allows the argument 'all'}

\item{all}{Returns a data frame of all early day motions}
}
\description{
Imports data on early day motions
}
\examples{
\dontrun{
x <- early_day_motions('all')
}
}
\keyword{EDM}

