% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_results.R
\name{election_results}
\alias{election_results}
\title{Elections Results}
\usage{
election_results(resultType = c("all", "ID"))
}
\arguments{
\item{resultType}{Accepts the arguments "all" and "ID"}

\item{all}{Returns general and by-election resuls for each consituency from the 2010 general election onwards.}

\item{ID}{Returns general and by-election resuls for each consituency from the 2010 general election onwards.}
}
\description{
Imports data on general election results
}
\examples{
\dontrun{
x <- election_results("all")

x <- election_results("ID")
}
}
\keyword{Election}
\keyword{Results}

