% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_results.R
\name{election_results}
\alias{election_results}
\title{election_results}
\usage{
election_results(ID = NULL)
}
\arguments{
\item{ID}{Accepts an ID for a general or by-election from the 2010 general election onwards, and returns the results. If NULL, returns all available election results. Defaults to NULL.}
}
\description{
Imports data on general election results
}
\examples{
\dontrun{
x <- election_results(ID=382037)
}
}
\keyword{Election}
\keyword{Results}
