% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\name{elections}
\alias{elections}
\title{Elections}
\usage{
elections(electType = c("all", "ID"))
}
\arguments{
\item{electType}{Allows the arguments "all" and "ID"}

\item{all}{Returns a data frame with the date and type of all general and by-elections since 1945.}

\item{ID}{Requests an election ID, and returns a data frame with the date and type of that election.}
}
\description{
Imports data on elections
}
\examples{
\dontrun{
x <- elections("all")

x <- elections("ID")
}
}
\keyword{Elections}

