% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\name{elections}
\alias{elections}
\title{Imports data on elections}
\usage{
elections(ID = NULL, extra_args = NULL, tidy = TRUE)
}
\arguments{
\item{ID}{Accepts an ID for a general or by-election from the 2010 general election onwards, and returns the date and type of the elction. If NULL, returns the date and type of all available elections. Defaults to NULL.}

\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}

\item{tidy}{Fix the variable names in the tibble to remove extra characters, superfluous text and convert variable names to snake_case. Defaults to TRUE.}
}
\value{
A tibble with details on all elections from the 2010 general election onwards, subject to function parameters.
}
\description{
Imports data on elections
}
\examples{
\dontrun{

x <- elections(517994)
}
}
\keyword{Elections}
