% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election.R
\name{elections}
\alias{elections}
\title{Elections}
\usage{
elections(all = TRUE)
}
\arguments{
\item{all}{Imports all available elections Defaults to TRUE.}
}
\description{
This imports data on elections
}
\examples{
\donttest{
### NOT RUN:
# x <- elections(all = TRUE)
# Returns a data frame with the date and type of all general and by-elections since 1945
}
}
\keyword{Elections}

