% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\name{elections}
\alias{elections}
\title{elections}
\usage{
elections(ID = NULL)
}
\arguments{
\item{ID}{Accepts an ID for a general or by-election from the 2010 general election onwards, and returns the date and type of the elction. If NULL, returns the date and type of all available elections. Defaults to NULL.}
}
\description{
Imports data on elections
}
\examples{
\dontrun{

x <- elections(517994)
}
}
\keyword{Elections}
