% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epetition.R
\name{epetition}
\alias{epetition}
\title{ePetitions}
\usage{
epetition(petType = c("all", "byConstituency", "petitionID", "response",
  "idConstituency"))
}
\arguments{
\item{petType}{The type of data you want, allows the arguments 'all', 'byConstituency',
# 'petitionID', 'response', and 'idConstituency'}

\item{all}{Imports all available epetitions}

\item{byConstituency}{Imports all available epetition signatures by constituency}

\item{petitionID}{Requests a petition ID, and then returns information on that petition}

\item{response}{Requests a petition ID, and then returns the governments response to that petition}

\item{idConstituency}{Requests a petition ID, and then returns the signatures per constituency for that petition}
}
\description{
Imports data on petitions
}
\examples{
\donttest{
#### NOT RUN
# x <- epetition("all")

# x <- epetition("byConstituency")

# x <- epetition("petitionID")
### RETURNS:
# Enter the petition ID:

# x <- epetition("response")
### RETURNS:
# Enter the petition ID:

# x <- epetition("idConstituency")
### RETURNS:
# Enter the petition ID:

}
}
\keyword{ePetitions}

