\name{hansard}
\alias{hansard}
\title{Hansard}
\usage{
oral_questions()
written_questions()
petitions()
commons_terms()
commons_divisions()
tv_programmes()
commons_bills()
commons_answered_questions()
oral_question_times()
constituencies()
early_day_motions()
election_results()
elections()
lords_attendance()
lords_ammendments()
lords_divisions()
lords_written_questions()
all_members()
commons_members()
lords_members()
commons_interest()
lords_interest()
papers_laid()
all_answered_questions()
publication_logs()
research_briefings()
sessions()
commons_terms()
petitions()
}
\description{
Retrieves data from the data.parliament.uk API. Despite the name, the API does not current give access to the Hansard.
}
\examples{
oral_questions()
written_questions()
petitions()
commons_terms()
commons_divisions()
tv_programmes()
commons_bills()
commons_answered_questions()
oral_question_times()
constituencies()
early_day_motions()
election_results()
elections()
lords_attendance()
lords_ammendments()
lords_divisions()
lords_written_questions()
all_members()
commons_members()
lords_members()
commons_interest()
lords_interest()
papers_laid()
all_answered_questions()
publication_logs()
research_briefings()
sessions()
commons_terms()
petitions()
}
