% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard-package.R
\docType{package}
\name{hansard}
\alias{hansard}
\alias{hansard-package}
\title{hansard: Provides Easy Downloading Capabilities for the UK Parliament API}
\description{
Provides functions to request data from the data.parliament.uk APIs. Because of the structure of the data.parliament.uk API, there is a named function for each type of available data for ease of use. Functions for each new API will be added as and when they become available on data.parliament.uk. The API is rate limited to returning 5500 rows per request in some instances, although in testing this has been inconsistent.
}
\details{
The API is still in beta, and only about half of the planned datasets are currently available. The package name is ironic and hopeful, as the actual contents of the Hansard are not yet available through this API.

In addition to the more generic function names, each function in hansard has a wrapper where the name is prefixed with \code{'hansard_'}. For example, both \code{bills()} and \code{hansard_bills()} will return the same result.

For more details please see the \href{https://cran.r-project.org/web/packages/hansard/vignettes/introduction.html}{vignette}, or the API documentation on \url{http://explore.data.parliament.uk/}.
}
