% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_basic.R
\name{hansard_basic}
\alias{hansard_basic}
\title{hansard_basic}
\usage{
hansard_basic()
}
\description{
Walks the user through the various steps of calling data from the parliamentary API.
}
\examples{
\dontrun{
x <- hansard_basic()
}
}
\keyword{API}
\keyword{Hansard}
