% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_generic.R
\name{hansard_generic}
\alias{hansard_generic}
\title{Hansard - Generic API Function}
\usage{
hansard_generic(path)
}
\arguments{
\item{path}{The url path to the data you wish to retrieve}
}
\description{
A semi-generic function for the parliamentary API. Provides greater flexibility, including any newly released features or data not yet included in the individual functions of the hansard package. Users must specify '.json' in their path.
}
\examples{
\dontrun{
x <- hansard_generic('elections.json')
}
}
\keyword{API}
\keyword{Hansard}

