% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_amendments.R
\name{lords_amendments}
\alias{lords_amendments}
\title{lords_amendments}
\usage{
lords_amendments(all = TRUE)
}
\arguments{
\item{all}{Returns a data frame with all available House of Lords Amendments. Defaults to TRUE.}
}
\description{
Imports data on House of Lords Amendments
}
\examples{
\dontrun{
x <- lords_amendments()
}
}
\keyword{Amendments}
\keyword{House}
\keyword{Lords}
\keyword{of}
