% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_ammendments.R
\name{lords_ammendments}
\alias{lords_ammendments}
\title{House of Lords Ammendments}
\usage{
lords_ammendments(all = TRUE)
}
\arguments{
\item{all}{Returns a data frame with all available House of Lords Ammendments. Defaults to TRUE.}
}
\description{
Imports data on House of Lords Ammendments
}
\examples{
\dontrun{
x <- lords_ammendments()
}
}
\keyword{Ammendments}
\keyword{House}
\keyword{Lords}
\keyword{of}

