% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_attendance.R
\name{lords_attendance}
\alias{lords_attendance}
\title{House of Lords Attendance}
\usage{
lords_attendance(type = c("all", "date"))
}
\arguments{
\item{type}{Accepts arguments "all" and "date"}

\item{all}{Imports all available House of Lords attendance records}

\item{date}{all available House of Lords attendance records on a given date}
}
\description{
This imports data on House of Lords attendance
}
\examples{
lords_attendance()
}
\keyword{Attendance}
\keyword{House}
\keyword{Lords}
\keyword{of}

