% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_attendance.R
\name{lords_attendance}
\alias{lords_attendance}
\title{lords_attendance}
\usage{
lords_attendance(lordsAttendType = c("all", "date"))
}
\arguments{
\item{lordsAttendType}{Accepts arguments 'all' and 'date'.}

\item{all}{Returns a data frame with all available House of Lords attendance records.}

\item{date}{Requests a date and returns a data frame with all available House of Lords attendance records for that date.}
}
\description{
Imports data on House of Lords attendance
}
\examples{
\dontrun{
x <- lords_attendance('all')

x <- lords_attendance('date')
}

}
\keyword{Attendance}
\keyword{House}
\keyword{Lords}
\keyword{of}

