% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_attendance.R
\name{lords_attendance}
\alias{lords_attendance}
\title{House of Lords Attendance}
\usage{
lords_attendance(lordsAttendType = c("all", "date"))
}
\arguments{
\item{lordsAttendType}{Accepts arguments 'all' and 'date'}

\item{all}{Imports all available House of Lords attendance records}

\item{date}{Imports all available House of Lords attendance records for a given date}
}
\description{
This imports data on House of Lords attendance
}
\examples{
\donttest{
#### NOT RUN:
# x <- lords_attendance('all')
#Returns all House of Lords attendance

# x <- lords_attendance('date')
#Returns House of Lords attendance data for a given date }
}
\keyword{Attendance}
\keyword{House}
\keyword{Lords}
\keyword{of}

