% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_divisions.R
\name{lords_divisions}
\alias{lords_divisions}
\title{House of Lords Divisions}
\usage{
lords_divisions(type = c("all", "date", "notContent", "content"))
}
\arguments{
\item{all}{Imports all available divisions.}

\item{date}{Imports all available divisions on a date.}

\item{notContent}{Imports all divisions where a given lord voted not content.}

\item{content}{Imports all divisions where a given lord voted content.}
}
\description{
This imports data on House of Lords divisions
#' @param type The type of data you want, allows the arguments "all", "date", "no", "aye"
}
\examples{
x <- lords_divisions("all")


x <- lords_divisions("date")


x <- lords_divisions("notContent")


x <- lords_divisions("content")


}
\keyword{Divisions}
\keyword{Lords}

