% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_divisions.R
\name{lords_divisions}
\alias{lords_divisions}
\title{House of Lords Divisions}
\usage{
lords_divisions(lordsDivType = c("all", "date", "notContent", "content"))
}
\arguments{
\item{lordsDivType}{Accepts arguments 'all', 'date', 'notContent' and 'content'}

\item{all}{Imports all available divisions.}

\item{date}{Imports all available divisions on a date.}

\item{notContent}{Imports all divisions where a given lord voted not content.}

\item{content}{Imports all divisions where a given lord voted content.}
}
\description{
This imports data on House of Lords divisions
#' @param type The type of data you want, allows the arguments 'all', 'date', 'no', 'aye'
}
\examples{
\donttest{
x <- lords_divisions('all')
# Returns all divisions in the House of Lords

x <- lords_divisions('date')
# Requests a date, and then returns all divisions on that date

x <- lords_divisions('notContent')
# Requests a member ID, and then returns all divisions in which they have voted not content

x <- lords_divisions('content')
# Requests a member ID, and then returns all divisions in which they have voted content
}
}
\keyword{Divisions}
\keyword{Lords}

