% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_divisions.R
\name{lords_divisions}
\alias{lords_divisions}
\title{House of Lords Divisions}
\usage{
lords_divisions(lordsDivType = c("all", "date", "notContent", "content",
  "session"))
}
\arguments{
\item{lordsDivType}{Accepts arguments 'all', 'date', 'notContent' and 'content'}

\item{all}{Returns a data frame with all available divisions in the House of Lords.}

\item{date}{Requests a date, and then returns a data frame with all divisions on that date.}

\item{notContent}{Requests a member ID, and then returns a data frame with all divisions in which they have voted not content.}

\item{content}{Requests a member ID, and then returns a data frame with all divisions in which they have voted content.}
}
\description{
Imports data on House of Lords divisions
#' @param type The type of data you want, allows the arguments 'all', 'date', 'no', 'aye'
}
\examples{
\dontrun{
x <- lords_divisions('all')

x <- lords_divisions('date')

x <- lords_divisions('notContent')

x <- lords_divisions('content')
}
}
\keyword{Divisions}
\keyword{Lords}

