% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_vote_record.R
\name{lords_vote_record}
\alias{lords_vote_record}
\title{lords_vote_record}
\usage{
lords_vote_record(lord.id, lordsRecord = c("all", "content", "notContent"))
}
\arguments{
\item{lord.id}{The ID number of a member of the House of Lords. To look up the ID number of a member of the House of Lords use the members_search() function.}

\item{lordsRecord}{Accepts the arguments 'all', 'content' and 'notContent'. Defaults to 'all'.}

\item{all}{Returns a data frame with all recorded votes for a given member of the house of lords.}

\item{content}{Returns a data frame with all divisions where a given lord voted aye.}

\item{notContent}{Returns a data frame with all divisions where a given lord voted no.}
}
\description{
Accepts an ID number for a member of the House of Lords, and returns a data frame of their votes. Provides similar functionality to the lords_divisions() function, but accepts member IDs as function parameters rather than requesting them from the console.
}
\examples{
\dontrun{
x <- lords_vote_record(172, 'all')
}

}
\keyword{divisions}
