% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_written_questions.R
\name{lords_written_questions}
\alias{lords_written_questions}
\title{House of Lords Attendance}
\usage{
lords_written_questions(all = TRUE)
}
\arguments{
\item{all}{Imports all available House of Lords written questions Defaults to TRUE.}
}
\description{
This imports data on House of Lords written questions
}
\examples{
x <- lords_written_questions()
}
\keyword{House}
\keyword{Lords}
\keyword{Questions}
\keyword{Written}
\keyword{of}

