% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_written_questions.R
\name{lords_written_questions}
\alias{lords_written_questions}
\title{House of Lords Written Questions}
\usage{
lords_written_questions(lordsWritType = c("all", "department", "dates"))
}
\arguments{
\item{lordsWritType}{The type of data you want, allows the arguments 'all', 'department' and 'dates'}

\item{all}{Imports all available written questions}

\item{department}{Imports all available written questions answered by a given department}

\item{dates}{Imports all available written questions from between two given dates}
}
\description{
Imports data on House of Lords written questions
}
\examples{
\donttest{
# NOT RUN
# Returns all written questions from the house of lords
# x <- lords_written_questions('all')
# head(x)

# NOT RUN
# Requests a department, and then returns all written questions answered by that department
# x <- lords_written_questions('department')
# Enter the name of the answering department: Health
# head(x) # Returns all answered questions where the answering department
          # includes 'Health' in its name
# NOT RUN
# Requests two dates, and then returns all written questions within those two dates
# x <- lords_written_questions('dates')
# x <- lords_written_questions('dates')
####RETURNS:
# Enter start date(yyyy-mm-dd): 2016-10-10
# Enter start date(yyyy-mm-dd): 2016-10-15
# head(x) }
}
\keyword{House}
\keyword{Lords}
\keyword{Questions}
\keyword{Written}
\keyword{of}

