% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lords_written_questions.R
\name{lords_written_questions}
\alias{lords_written_questions}
\title{lords_written_questions}
\usage{
lords_written_questions(lordsWritType = c("all", "department", "dates"))
}
\arguments{
\item{lordsWritType}{The type of data you want, allows the arguments 'all', 'department' and 'dates'}

\item{all}{Returns a data frame all written questions from the house of lords}

\item{department}{Requests a department, and then returns a data frame with all written questions answered by that department}

\item{dates}{Requests two dates, and returns a data frame with all available written questions from between the two given dates}
}
\description{
Imports data on House of Lords written questions
}
\examples{
\dontrun{
x <- lords_written_questions('all')

# x <- lords_written_questions('department')

# x <- lords_written_questions('dates')
}
}
\keyword{House}
\keyword{Lords}
\keyword{Questions}
\keyword{Written}
\keyword{of}

