% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members}
\alias{members}
\title{All Members of Parliament}
\usage{
members(house = c("all", "commons", "lords", "lordsInterests"))
}
\arguments{
\item{house}{The type of data you want, allows the arguments}

\item{all}{Imports all available members. The default value return}

\item{commons}{Imports all available members}

\item{lords}{Imports all available members}

\item{lordsInterests}{Imports all available members}
}
\description{
Imports data on All Members of Parliament including the Lords and the Commons
}
\examples{
\donttest{
### NOT RUN:
# x <- members('all')
# Returns a data frame with information on all members of Parliament, including both
# the House of Lords and the House of Commons. The data frame includes both current and
# previous members, and the API currently does not have information on when a member first
# sat in the house, or to distinguish current from former members.

# y <- members('commons')
# Returns a data frame with information on all members of the House of Commons. The data
# frame includes both current and previous members of the House of Commons, and the API
# currently does not have information on when a member first sat in the house, or to
# distinguish current from former members.

# z <- members('lords')
# Returns a data frame with information on all members of the House of Lords. The data
# frame includes both current and previous members of the House of Lords, and the API
# currently does not have information on when a member first sat in the house, or to
# distinguish current from former members.


# b <- members('lordsInterests')
# Requests a member ID, and returns a dataframe of the registered interests of that
# member of the House of Lords. }
}
\keyword{All}
\keyword{Members}
\keyword{Parliament}
\keyword{of}

