% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members}
\alias{members}
\alias{hansard_members}
\alias{commons_members}
\alias{hansard_commons_members}
\alias{lords_members}
\alias{hansard_lords_members}
\title{Members of Both Houses}
\usage{
members(ID = NULL, extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)

hansard_members(ID = NULL, extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)

commons_members(extra_args = NULL, tidy = TRUE, tidy_style = "snake_case",
  verbose = FALSE)

hansard_commons_members(extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)

lords_members(extra_args = NULL, tidy = TRUE, tidy_style = "snake_case",
  verbose = FALSE)

hansard_lords_members(extra_args = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)
}
\arguments{
\item{ID}{The ID of a member of the House of Commons or the House of Lords. Defaults to \code{NULL}. If \code{NULL}, returns a tibble of all members. If not NULL, returns a tibble with basic information on that member.}

\item{extra_args}{Additional parameters to pass to API. Defaults to \code{NULL}.}

\item{tidy}{Fix the variable names in the tibble to remove special characters and superfluous text, and converts the variable names to a consistent style. Defaults to \code{TRUE}.}

\item{tidy_style}{The style to convert variable names to, if \code{tidy = TRUE}. Accepts one of 'snake_case', 'camelCase' and 'period.case'. Defaults to 'snake_case'.}

\item{verbose}{If \code{TRUE}, returns data to console on the progress of the API request. Defaults to \code{FALSE}.}
}
\value{
A tibble with data on members of the House of Commons and/or the House of Lords.
}
\description{
Imports data on all current and former Members of Parliament including the Lords and the Commons
}
\section{Member details functions}{

\describe{
\item{members}{Basic details on a given member}
\item{commons_members}{MPs in the House of Commons}
\item{lords_members}{Peers in the House of Lords}
}
}

\examples{
\dontrun{
x <- members(172)

x <- commons_members()

x <- lords_members()

}
}
\seealso{
\code{\link{members_search}}
}
