% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members}
\alias{members}
\alias{commons_members}
\alias{lords_members}
\alias{lords_interests}
\title{members}
\usage{
members(ID = NULL, extra_args = NULL, tidy = TRUE)

commons_members(extra_args = NULL, tidy = TRUE)

lords_members(extra_args = NULL, tidy = TRUE)

lords_interests(peer_id = NULL, extra_args = NULL, tidy = TRUE)
}
\arguments{
\item{ID}{The ID of a member of the House of Commons or the House of Lords. Defaults to NULL. If NULL, returns a tibble of all members. If not NULL, returns a tibble with basic information on that member.}

\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}

\item{tidy}{Fix the variable names in the tibble to remove extra characters, superfluous text and convert variable names to snake_case. Defaults to TRUE.}

\item{peer_id}{The ID of a member of the house of lords. If NULL, returns a tibble with all listed financial interests for all members. Defaults to NULL.}
}
\description{
Imports data on all current and former Members of Parliament including the Lords and the Commons

Imports data on all current and former MPs

Imports data on all current and former peers
}
\examples{
\dontrun{
x <- members(172)

x <- commons_members()

x <- lords_members()

x <- lords_interests(530)
}
}
\keyword{All}
\keyword{Members}
\keyword{Parliament}
\keyword{of}
