% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members}
\alias{members}
\alias{commons_members}
\alias{lords_members}
\alias{commons_interests}
\alias{lords_interests}
\title{members}
\usage{
members(ID = NULL, extra_args = NULL)

commons_members(extra_args = NULL)

lords_members(extra_args = NULL)

commons_interests(extra_args = NULL)

lords_interests(peer_id = NULL, extra_args = NULL)
}
\arguments{
\item{ID}{The ID of a member of the House of Commons or the House of Lords. Defaults to NULL. If NULL, returns a data frame of all members. If not NULL, returns a data frame with basic information on that member.}

\item{extra_args}{Additional parameters to pass to API. Defaults to NULL.}

\item{peer_id}{The ID of a member of the house of lords. If NULL, returns a data frame with all listed financial interests for all members. Defaults to NULL.}
}
\description{
Imports data on All Members of Parliament including the Lords and the Commons

commons_members
}
\examples{
\dontrun{
x <- members(172)

x <- commons_members()

x <- lords_members()

x <- commons_interests()

x <- lords_interests(530)
}
}
\keyword{All}
\keyword{Members}
\keyword{Parliament}
\keyword{of}
