% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members_search.R
\name{members_search}
\alias{members_search}
\alias{hansard_members_search}
\title{Search for an MP or Peer by name and constituency}
\usage{
members_search(search = NULL, tidy = TRUE, tidy_style = "snake_case",
  verbose = FALSE)

hansard_members_search(search = NULL, tidy = TRUE,
  tidy_style = "snake_case", verbose = FALSE)
}
\arguments{
\item{search}{Accepts any string. Defaults to \code{NULL}. If \code{NULL}, returns a tibble with all members of both houses of parliament.}

\item{tidy}{Fix the variable names in the tibble to remove extra characters, superfluous text and convert variable names to snake_case. For the `members_search` function it also changes the '_about' column name to 'mnis_id' (or 'mnisId' or 'mnis.id', depending on the value of the `tidy_text` parameter, and removes the URL to preserve only the numerical ID. Defaults to \code{TRUE}.}

\item{tidy_style}{The style to convert variable names to, if \code{tidy = TRUE}. Accepts one of \code{'snake_case'}, \code{'camelCase'} and \code{'period.case'}. Defaults to \code{'snake_case'}.}

\item{verbose}{If \code{TRUE}, returns data to console on the progress of the API request. Defaults to \code{FALSE}.}
}
\value{
A tibble with the results of the search.
}
\description{
Function searches for the string and returns a tibble with all matches from both houses of parliament. Returns all partial matches in the members' names, constituencies, twitter handle and webpage. The default search is NULL, which returns a tibble of all members of both houses, the same result as \code{members('all')}.
}
\examples{
\dontrun{
x <- members_search('chris')

x <- members_search(search='chris')
}
}
\seealso{
\code{\link{members}}
}
