% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members_search}
\alias{members_search}
\title{MP Search}
\usage{
members_search(Search = NULL)
}
\arguments{
\item{all}{Imports all available members Defaults to TRUE.}
}
\description{
This imports data on All Members of Parliament including the Lords and the Commons
}
\examples{
# Function searches for the string and returns a data frame with all matches from both houses of parliament. Returns all partial matches in the members' names, constituencies, twitter handle and webpage.
members_search ("query")







}
\keyword{All}
\keyword{Members}
\keyword{Parliament}
\keyword{of}

