% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_edms.R
\name{mp_edms}
\alias{mp_edms}
\title{mp_edms}
\usage{
mp_edms(mp.id, edmType = c("signed", "sponsor"))
}
\arguments{
\item{mp.id}{Required parameter. The ID number of an MP.}

\item{edmType}{The type of data you want, allows the arguments "signed" and "sponsor". Defaults to "signed".}

\item{signed}{Returns a data frame of all early day motions signed by the given MP.}

\item{Sponsor}{Returns a data frame of early day motions where the given member is the primary sponsor or a sponsor.}
}
\description{
Imports data on early day motions
}
\examples{
\dontrun{
x <- mp_edms('signed')

x <- mp_edms('sponsor')

}
}
\keyword{Day}
\keyword{Early}
\keyword{Motion}

