% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons_divisions.R
\name{mp_no}
\alias{mp_no}
\title{Divisions where an MP voted no}
\usage{
mp_no(mp.id)
}
\arguments{
\item{mp.id}{Accepts the ID number of a member of the House of Commons}
}
\description{
Accepts an ID number for a member of the house of commons, and returns a data frame of all divisions where they voted aye.
}
\examples{
\dontrun{
x <- mp_no(172)
}
}
\keyword{divisions}

