% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_questions.R
\name{mp_questions}
\alias{mp_questions}
\alias{hansard_mp_questions}
\title{Oral and written questions asked by MPs}
\usage{
mp_questions(mp_id = NULL, question_type = "all",
  start_date = "1900-01-01", end_date = Sys.Date(), extra_args = NULL,
  tidy = TRUE, tidy_style = "snake_case", verbose = FALSE)

hansard_mp_questions(mp_id = NULL, question_type = "all",
  start_date = "1900-01-01", end_date = Sys.Date(), extra_args = NULL,
  tidy = TRUE, tidy_style = "snake_case", verbose = FALSE)
}
\arguments{
\item{mp_id}{The ID number of a member of the House of Commons. Defaults to \code{NULL}.}

\item{question_type}{Accepts the arguments \code{'all'}, \code{'oral'} and \code{'written'}. Defaults to \code{'all'}.}

\item{start_date}{The earliest date to include in the tibble. Accepts character values in \code{'YYYY-MM-DD'} format, and objects of class \code{Date}, \code{POSIXt}, \code{POSIXct}, \code{POSIXlt} or anything else than can be coerced to a date with \code{as.Date()}. Defaults to \code{'1900-01-01'}.}

\item{end_date}{The latest date to include in the tibble. Defaults to \code{'1900-01-01'}. Accepts character values in \code{'YYYY-MM-DD'} format, and objects of class \code{Date}, \code{POSIXt}, \code{POSIXct}, \code{POSIXlt} or anything else than can be coerced to a date with \code{as.Date()}. Defaults to the current system date.}

\item{extra_args}{Additional parameters to pass to API. Defaults to \code{NULL}.}

\item{tidy}{Fix the variable names in the tibble to remove special characters and superfluous text, and converts the variable names to a consistent style. Defaults to \code{TRUE}.}

\item{tidy_style}{The style to convert variable names to, if \code{tidy = TRUE}. Accepts one of \code{'snake_case'}, \code{'camelCase'} and \code{'period.case'}. Defaults to \code{'snake_case'}.}

\item{verbose}{If \code{TRUE}, returns data to console on the progress of the API request. Defaults to \code{FALSE}.}
}
\value{
A tibble with details on all questions asked by a member of the House of Commons.
}
\description{
Accepts an ID number for a member of the House of Commons, and returns a tibble of of all their oral and written questions.
}
\examples{
\dontrun{
x <- mp_questions(172, 'all')
}

}
\seealso{
\code{\link{all_answered_questions}}

\code{\link{commons_answered_questions}}

\code{\link{commons_oral_questions}}

\code{\link{commons_oral_question_times}}

\code{\link{commons_written_questions}}

\code{\link{lords_written_questions}}
}
