% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_questions.R
\name{mp_questions}
\alias{mp_questions}
\title{mp_questions}
\usage{
mp_questions(mp.id, questionType = c("all", "oral", "written"))
}
\arguments{
\item{mp.id}{The ID number of a member of the House of Commons.}

\item{questionType}{Accepts the arguments 'all', 'oral' and 'written'. Defaults to 'all'.}

\item{all}{Returns a data frame with all recorded votes for a given member of the house of commons.}

\item{oral}{Returns a data frame with all divisions where a given MP voted aye.}

\item{written}{Returns a data frame with all divisions where a given MP voted no.}
}
\description{
Accepts an ID number for a member of the house of commons, and returns a data frame of of all their votes.
}
\examples{
\dontrun{
x <- mp_questions(172, 'all')
}

}
\keyword{divisions}

