% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_vote_record.R
\name{mp_vote_record}
\alias{mp_vote_record}
\title{mp_vote_record}
\usage{
mp_vote_record(mp_id = NULL, date = NULL, lobby = c("all", "aye", "no"))
}
\arguments{
\item{mp_id}{The ID number of a member of the House of Commons. To look up the ID number of a member of the House of Commons use the members_search() function.}

\item{date}{Returns all divisions on a given date. Defaults to NULL.}

\item{lobby}{Accepts one of 'all', 'aye', 'no'. 'aye' returns votes where the MP voted 'aye', 'no' returns votes where the MP voted 'no', 'all' returns all available votes by the MP. Defaults to 'all'.}
}
\description{
Accepts an ID number for a member of the House of Commons, and returns a data frame of their votes. Provides similar functionality to the commons_divisions() function, but accepts member IDs as function parameters rather than requesting them from the console.
}
\examples{
\dontrun{
x <- mp_vote_record(172, lobby='all')

x <- mp_vote_record(172, lobby='aye')

x <- mp_vote_record(172, lobby='no')
}
}
\keyword{divisions}
