% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papers_laid.R
\name{papers_laid}
\alias{papers_laid}
\title{Papers Laid}
\usage{
papers_laid(all = TRUE)
}
\arguments{
\item{all}{Imports all available papers laid Defaults to TRUE.}
}
\description{
This imports data on Papers Laid
}
\examples{
papers_laid
}
\keyword{Laid}
\keyword{Papers}

