% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papers_laid.R
\name{papers_laid}
\alias{papers_laid}
\title{Papers Laid}
\usage{
papers_laid(paperType = c("all"))
}
\arguments{
\item{paperType}{The type of data you want, allows the arguments 'all', 'department' and 'dates'}

\item{all}{Imports all available papers laid}
}
\description{
Imports data on Papers Laid
}
\examples{
\donttest{
#### NOT RUN:
# x <- papers_laid('all') }

}
\keyword{Laid}
\keyword{Papers}

