% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papers_laid.R
\name{papers_laid}
\alias{papers_laid}
\title{papers_laid}
\usage{
papers_laid(all)
}
\arguments{
\item{all}{Returns a data frame with all available papers laid.}
}
\description{
Imports data on Papers Laid
}
\examples{
\dontrun{
x <- papers_laid('all') }

}
\keyword{Laid}
\keyword{Papers}
