% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_logs.R
\name{publication_logs}
\alias{publication_logs}
\title{publication_logs}
\usage{
publication_logs(all = TRUE)
}
\arguments{
\item{all}{Returns a data frame with all available Publication Logs. Defaults to TRUE.}
}
\description{
Imports data on Publication Logs
}
\examples{
\dontrun{
# x <- publication_logs()
}
}
\keyword{Logs}
\keyword{Publication}
