% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_logs.R
\name{publication_logs}
\alias{publication_logs}
\title{publication_logs}
\usage{
publication_logs(ID = NULL, start_date = "1900-01-01",
  end_date = Sys.Date())
}
\arguments{
\item{ID}{Publication ID. Defaults to NULL. If not null, requests a data frame with information on the given publication.}

\item{start_date}{The earliest date to include in the data frame. Defaults to '1900-01-01'.}

\item{end_date}{The latest date to include in the data frame. Defaults to current system date.}
}
\description{
Imports data on Publication Logs
}
\examples{
\dontrun{
# x <- publication_logs(683267)
}
}
\keyword{Logs}
\keyword{Publication}
