% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_logs.R
\name{publication_logs}
\alias{publication_logs}
\title{Publication Logs}
\usage{
publication_logs(all = TRUE)
}
\arguments{
\item{all}{Imports all available Publication Logs. Defaults to TRUE.}
}
\description{
This imports data on Publication Logs
}
\examples{
x <- publication_logs(all = TRUE)
# returns a data frame with all publication logs
# NOT RUN:
# x <- publication_logs()
# head(x) ###
}
\keyword{Logs}
\keyword{Publication}

