% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_briefings.R
\name{research_briefings}
\alias{research_briefings}
\title{Parliamentary Research Briefings}
\usage{
research_briefings(resType = c("all", "topics", "types", "byTopic",
  "subTopic", "topicSubTopic"))
}
\arguments{
\item{resType}{The type of data you want, allows the arguments 'all', 'topics', 'types',
# 'byTopic','subTopic' and 'topicSubTopic'}

\item{all}{Imports a data frame with all available Parliamentary Research Briefings.}

\item{topics}{Imports a data frame with all Parliamentary Research Briefings topics.}

\item{types}{Imports a data frame with all Parliamentary Research Briefings types.}

\item{subTopic}{Requests a topic, and imports a data frame with all of the possible
# sub-topics for that topic.}

\item{byTopic}{Requests a topic, and imports a data frame with all available
# Parliamentary Research Briefings on that topic}

\item{topicSubTopic}{Requests a topic and a subtopic, and imports a data frame with
# all available
# Parliamentary Research Briefings on that subtopic}
}
\description{
This imports data on  Parliamentary Research Briefings
}
\examples{
\donttest{
# NOT RUN
# x <- research_briefings('all')
# head(x)

# x <- research_briefings('topics')
# print(x)

# x <- research_briefings('types')
# print(x)

# x <- research_briefings('byTopic')
# RETURNS:
# [1]Sub-topics are case sensititve. To return list of sub-topics, enter yes.
# Enter sub-topic:  #yes
# RETURNS:
# [1] 'Agriculture, animals, food and rural affairs' 'Asylum, immigration and nationality'
# [3] 'Business, industry and consumers'             'Communities and families'
# [5] 'Crime, civil law, justice and rights'         'Culture, media and sport'
# [7] 'Defence'                                      'Economy and finance'
# [9] 'Education'                                    'Employment and training'
# [11] 'Energy and environment'                       'European Union'
# [13] 'Health services and medicine'                 'Housing and planning'
# [15] 'International affairs'                        'Parliament, government and politics'
# [17] 'Science and technology'                       'Social Security and pensions'
# [19] 'Social services'                              'Transport'
# Enter Topic. For ease of use, copy and paste the topic: Education
# head(x)

# x <- research_briefings('subTopic')
# RETURNS:
# [1]Sub-topics are case sensititve. To return list of sub-topics, enter yes.
# Enter sub-topic:  #yes
# RETURNS:
# [1] 'Agriculture, animals, food and rural affairs' 'Asylum, immigration and nationality'
# [3] 'Business, industry and consumers'             'Communities and families'
# [5] 'Crime, civil law, justice and rights'         'Culture, media and sport'
# [7] 'Defence'                                      'Economy and finance'
# [9] 'Education'                                    'Employment and training'
# [11] 'Energy and environment'                       'European Union'
# [13] 'Health services and medicine'                 'Housing and planning'
# [15] 'International affairs'                        'Parliament, government and politics'
# [17] 'Science and technology'                       'Social Security and pensions'
# [19] 'Social services'                              'Transport'
# Enter Topic. For ease of use, copy and paste the topic: Education
# head(x)

# x <- research_briefings('topicSubTopic')
# NOT RUN
# [1]Sub-topics are case sensititve. To return list of sub-topics, enter yes.
# Enter sub-topic: yes
# RETURNS:
# [1] 'Agriculture, animals, food and rural affairs' 'Asylum, immigration and nationality'
# [3] 'Business, industry and consumers'             'Communities and families'
# [5] 'Crime, civil law, justice and rights'         'Culture, media and sport'
# [7] 'Defence'                                      'Economy and finance'
# [9] 'Education'                                    'Employment and training'
# [11] 'Energy and environment'                       'European Union'
# [13] 'Health services and medicine'                 'Housing and planning'
# [15] 'International affairs'                        'Parliament, government and politics'
# [17] 'Science and technology'                       'Social Security and pensions'
# [19] 'Social services'                              'Transport'
# Enter Topic. For ease of use, copy and paste the topic:
# [1]Sub-topics are case sensititve. To return list of sub-topics, enter yes.
# Enter sub-topic:}
}
\keyword{Briefings}
\keyword{Parliamentary}
\keyword{Research}

