% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_briefings.R
\name{research_briefings}
\alias{research_briefings}
\alias{research_topics}
\alias{research_types}
\title{research_briefings}
\usage{
research_briefings(topic = NULL, subtopic = NULL, type = NULL)

research_topics(subtopics = TRUE)

research_types()
}
\arguments{
\item{topic}{The topic of the parliamentary briefing.}

\item{subtopic}{The subtopic of the parliamentary briefing.}

\item{type}{The type of research briefing.}

\item{subtopics}{If TRUE, returns full list with all topics and subtopics. If FALSE, only returns list of topics. Defaults to TRUE.}
}
\description{
Imports data on  Parliamentary Research Briefings. To see a list of possible topics and subtopics, call \code{\link{research_topics}}. To see a list of briefing types, call \code{\link{research_types_list}}.

Prints or assigns to an object a list of topics and subtopics on  Parliamentary Research Briefings.

Prints or assigns to an object a list types of Parliamentary Research Briefings.
}
\examples{
\dontrun{
x <- research_briefings('Housing and planning')

#Requests can be made using the `research_topics_list`
# and `research_subtopics_list` included in the data
x <- research_briefings(topic = hansard::research_topics_list[[7]])

x <- research_briefings(subtopic = hansard::research_subtopics_list[[7]][10])

# Requests for certain briefing types can also be made using 'research_types_list'.
x <- research_briefings(type = hansard::research_types_list[[3]])


}

research_topics()


research_types()




}
\seealso{
research_topics
}
\keyword{Briefings}
\keyword{Parliamentary}
\keyword{Research}
