% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_briefings_lists.R
\name{research_topics_list}
\alias{research_topics_list}
\alias{hansard_research_topics_list}
\alias{research_subtopics_list}
\alias{hansard_research_subtopics_list}
\alias{research_types_list}
\alias{hansard_research_types_list}
\title{Lists of research briefing topics, subtopics and types.}
\usage{
research_topics_list()

hansard_research_topics_list()

research_subtopics_list()

hansard_research_subtopics_list()

research_types_list()

hansard_research_types_list()
}
\value{
A list with the different research topics available.

A list with the different research subtopics available.

A list with the different types of research briefings.
}
\description{
These functions do not accept any arguments.

A list of Parliamentary Research Briefings subtopics, grouped by topic.

A list of types of Parliamentary Research Briefings.
}
\details{
A list of Parliamentary Research Briefings topics.
}
\examples{
 \dontrun{

research_topics_list <- research_topics_list()
}

 \dontrun{

research_subtopics_list <- research_subtopics_list()

}
 \dontrun{

research_types_list <- research_types_list()

}
}
\seealso{
\code{\link{research_briefings}}
}
\keyword{Briefings}
\keyword{Parliamentary}
\keyword{Research}
