% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_briefings_lists.R
\name{research_topics_list}
\alias{research_topics_list}
\alias{research_subtopics_list}
\alias{research_types_list}
\title{research_topics_list}
\usage{
research_topics_list()

research_subtopics_list()

research_types_list()
}
\description{
Prints or assigns to an object a list of Parliamentary Research Briefings topics.

Prints or assigns to an object a list of Parliamentary Research Briefings subtopics, grouped by topic.

Prints or assigns to an object a list of types of Parliamentary Research Briefings.
}
\examples{

research_topics_list <- research_topics_list()



research_subtopics_list <- research_subtopics_list()



research_types_list <- research_types_list()




}
\keyword{Briefings}
\keyword{Parliamentary}
\keyword{Research}
