% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{research_subtopics_list}
\alias{research_subtopics_list}
\title{List of parliamentary research briefing topics and subtopics}
\format{A nested list with twenty levels, one for each possible briefing topic. Each level contains a list with all subtopics for the parent topic.}
\source{
\url{http://explore.data.parliament.uk/}
}
\usage{
research_subtopics_list
}
\description{
List of parliamentary research briefing topics and subtopics
}
\keyword{datasets}
