% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_programmes.R
\name{tv_programmes}
\alias{tv_programmes}
\title{tv_programmes}
\usage{
tv_programmes(AVtype = c("TV", "clips"))
}
\arguments{
\item{AVtype}{The type of data you want, allows the arguments 'TV' and 'clips'}

\item{TV}{Returns a data frame with details on all TV broadcasts produced by the Westminster Parliament.}

\item{clips}{Returns a data frame with details on all video clips produced by the Westminster Parliament.}
}
\description{
Imports data on TV broadcasts
}
\examples{
\dontrun{

x <- tv_programmes('TV')


x <- tv_programmes('clips')
}
}
\keyword{TV}
