% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_programmes.R
\name{tv_programmes}
\alias{tv_programmes}
\alias{tv_clips}
\alias{tv_channels}
\title{tv_programmes}
\usage{
tv_programmes(legislature = NULL)

tv_clips(mp_id = NULL)

tv_channels()
}
\arguments{
\item{legislature}{Accepts one of either 'commons' or 'lords'. If NULL, returns all TV programmes for all chambers.}

\item{mp_id}{Accepts the ID of an MP or peer, and returns all clips featuring that MP or peer. If NULL, returns data on all available clips. Defaults to NULL.}
}
\description{
Imports data on TV broadcasts. To import information on TV channel options,

Imports data on TV broadcasts. To import information on TV channel options,

Imports data on TV broadcasts
}
\examples{
\dontrun{

x <- tv_programmes('commons')

}
\dontrun{
x <- tv_clips(4335)
}
}
\keyword{TV}
